Add a backend abstraction to the portmap plugin
